+++ src/main/java/hmim/eteam/rest/backend/entity/course/CourseTheme.java
@@ -37,6 +37,6 @@ public class CourseTheme extends OrderedEntity {
         return new Theme().
                 id(getId()).
                 name(getName()).
-                priority(getPriority().intValue());
+                priority(getPriority());
     }
 }
